requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
scipy>=1.7.0

# Deep Learning
torch>=1.10.0
torch-geometric>=2.0.0
torch-scatter>=2.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Utils
pyyaml>=5.4.0
tqdm>=4.62.0
joblib>=1.1.0

# Logging
loguru>=0.5.3

# Testing (optional)
pytest>=6.2.0
pytest-cov>=2.12.0

# Experiment tracking (optional)
# wandb>=0.12.0
# tensorboard>=2.7.0
